Moved registry functions to their own file
